+++ changelog.md
@@ -6,6 +6,9 @@ This change log records updates after 2025-3-24.
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
 ## [v0.9.0] - 2025-10-15 (UTC-7)
+### Added
+- Add support for GRU layers in the neural network module
+
 ### Changed
 - Improve handling of empty arrays in `sum_of_square_total` and `standard_deviation` functions, and optimize probability distribution normalization
 - Organize the pooling layer and the convolutional layer under their respective modules